@@ -5,7 +5,7 @@ version: 3.2.0
 publish_to: none
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -17,14 +17,14 @@ dependencies:
   ###########################
 
   smashlibs:
-    # path: ../../../SMASH/smashlibs/
-    git:
-      url: https://github.com/moovida/smashlibs.git
-      ref: e08e258
+    path: ./smashlibs/
+    # git:
+    #   url: https://github.com/moovida/smashlibs.git
+      # ref: e08e258
 
-  dart_hydrologis_utils: ^0.6.1
+  dart_hydrologis_utils: ^0.8.1
 
-  dart_jts: ^0.2.0
+  dart_jts: ^0.2.3
 
   ###########################
   # OTHERS
@@ -40,10 +40,10 @@ dependencies:
   after_layout: ^1.2.0
 
   # MIT
-  material_design_icons_flutter: ^5.0.6996
+  material_design_icons_flutter: ^7.0.7296
 
   # Apache 2.0
-  charts_flutter: ^0.12.0
+  # charts_flutter: ^0.12.0
 
   # MIT
   provider: ^6.0.3
@@ -52,16 +52,16 @@ dependencies:
   latlong2: ^0.8.1
 
   # MIT
-  flutter_tags: ^1.0.0-nullsafety.1
+  flutter_tags_x: ^1.1.0
 
   # BSD
-  flutter_map: ^2.2.0
+  flutter_map: ^4.0.0
     # git:
     #   url: https://github.com/moovida/flutter_map.git
     #   ref: version4smash
   
   # BSD
-  flutter_map_marker_cluster: ^0.5.4
+  flutter_map_marker_cluster: ^1.1.1
     # git:
     #   url: https://github.com/moovida/flutter_map_marker_cluster.git
     #   ref: version4smash
