@@ -7,15 +7,18 @@ import 'package:flutter_server/com/hydrologis/gss/libs/variables.dart';
 
 class SmashSession {
   /// Checks credentials and returns an error message or null if login is ok.
-  static Future<String> login(
-      String user, String password, Project project) async {
-    var responseText = await ServerApi.login(user, password, project.id);
+  static Future<String?> login(
+      String user, String password, WebProject project) async {
+    var responseText = await WebServerApi.login(user, password, project.id!);
     if (!responseText.startsWith(NETWORKERROR_PREFIX)) {
       var token = responseText;
       setSessionToken(token);
       setSessionUser(user, password, project);
 
-      var configMap = await ServerApi.getUserConfigurations();
+      var configMap = await WebServerApi.getUserConfigurations();
+      if (configMap == null) {
+        return "Could not retrieve user configurations.";
+      }
       var lastBasemap = configMap[KEY_BASEMAP];
       if (lastBasemap != null) {
         html.window.sessionStorage[KEY_BASEMAP] = lastBasemap;
@@ -43,12 +46,12 @@ class SmashSession {
 
   static List<String> getSessionUser() {
     return [
-      html.window.sessionStorage[KEY_USER],
-      html.window.sessionStorage[KEY_PWD]
+      html.window.sessionStorage[KEY_USER]!,
+      html.window.sessionStorage[KEY_PWD]!
     ];
   }
 
-  static void setSessionUser(String user, String pwd, Project project) {
+  static void setSessionUser(String user, String pwd, WebProject project) {
     html.window.sessionStorage[KEY_USER] = user;
     html.window.sessionStorage[KEY_PWD] = pwd;
     html.window.sessionStorage[KEY_PROJECT] = project.toJsonString();
@@ -58,7 +61,7 @@ class SmashSession {
     html.window.sessionStorage[KEY_TOKEN] = token;
   }
 
-  static String getSessionToken() {
+  static String? getSessionToken() {
     var token = html.window.sessionStorage[KEY_TOKEN];
     if (token == null) {
       html.window.location.reload();
@@ -66,13 +69,13 @@ class SmashSession {
     return token;
   }
 
-  static Project getSessionProject() {
+  static WebProject getSessionProject() {
     var projectJson = html.window.sessionStorage[KEY_PROJECT];
     if (projectJson == null) {
       html.window.sessionStorage.remove(KEY_TOKEN);
       html.window.location.reload();
     }
-    return Project.fromJson(projectJson);
+    return WebProject.fromJson(projectJson!);
   }
 
   static void setBasemap(String basemap) {
@@ -95,8 +98,8 @@ class SmashSession {
     html.window.sessionStorage[KEY_MAPCENTER] = "$lon;$lat;$zoom";
   }
 
-  static List<double> getMapcenter() {
-    String xyz = html.window.sessionStorage[KEY_MAPCENTER];
+  static List<double>? getMapcenter() {
+    String? xyz = html.window.sessionStorage[KEY_MAPCENTER];
     if (xyz == null) {
       return null;
     }
@@ -109,17 +112,17 @@ class SmashSession {
   }
 
   static void logout({mapCenter}) async {
-    var tokenheader = ServerApi.getTokenHeader();
+    var tokenheader = WebServerApi.getTokenHeader();
     var projectName = getSessionProject();
 
     html.window.sessionStorage.remove(KEY_USER);
     html.window.sessionStorage.remove(KEY_PWD);
     html.window.sessionStorage.remove(KEY_TOKEN);
-    String baseMap = html.window.sessionStorage.remove(KEY_BASEMAP);
+    String? baseMap = html.window.sessionStorage.remove(KEY_BASEMAP);
     html.window.sessionStorage.remove(KEY_MAPCENTER);
-    String bookmarks = html.window.sessionStorage.remove(KEY_BOOKMARKS);
+    String? bookmarks = html.window.sessionStorage.remove(KEY_BOOKMARKS);
 
-    await ServerApi.saveUserConfigurations(tokenheader, projectName,
+    await WebServerApi.saveUserConfigurations(tokenheader, projectName,
         basemap: baseMap, mapCenter: mapCenter, bookmarks: bookmarks);
   }
 }
