@@ -51,7 +51,7 @@ const PREVIOUSID = "previous";
 const DATAID = "dataid";
 const DATA = "data";
 const IMAGEDATA = "imagedata";
-const THUMBNAIL = "thumbnail";
+// const THUMBNAIL = "thumbnail";
 const LOGS = "logs";
 const NOTES = "notes";
 const TEXT = "text";
